+++ tubeguardian-bom/tg-common/src/main/java/io/tubeguardian/common/exception/TubeGuardianException.java
@@ -1,0 +1,12 @@
+package io.tubeguardian.common.exception;
+
+public abstract class TubeGuardianException extends RuntimeException {
+
+    protected TubeGuardianException(String message) {
+        super(message);
+    }
+
+    protected TubeGuardianException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
