--- tubeguardian-bom/tg-common/src/main/java/it/tubeguardian/common/exception/TubeGuardianException.java
@@ -1,12 +1,0 @@
-package it.tubeguardian.common.exception;
-
-public abstract class TubeGuardianException extends RuntimeException {
-
-    protected TubeGuardianException(String message) {
-        super(message);
-    }
-
-    protected TubeGuardianException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
