--- tubeguardian-bom/tg-common/src/main/java/it/tubeguardian/common/repository/BrandProfileRepository.java
@@ -1,9 +1,0 @@
-package it.tubeguardian.common.repository;
-
-import it.tubeguardian.common.domain.BrandProfile;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface BrandProfileRepository extends JpaRepository<BrandProfile, String> {
-}
