--- tubeguardian-bom/tg-common/src/main/java/it/tubeguardian/common/repository/VideoRepository.java
@@ -1,10 +1,0 @@
-package it.tubeguardian.common.repository;
-
-import it.tubeguardian.common.domain.Video;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.UUID;
-
-@Repository
-public interface VideoRepository extends JpaRepository<Video, UUID> { }
