+++ tubeguardian-bom/tg-common/src/test/java/io/tubeguardian/common/TestApplication.java
@@ -1,0 +1,7 @@
+package io.tubeguardian.common;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TestApplication {
+}
