--- tubeguardian-bom/tg-common/src/test/java/it/tubeguardian/common/TestApplication.java
@@ -1,7 +1,0 @@
-package it.tubeguardian.common;
-
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TestApplication {
-}
